ISPN-11494 Refactor delete cache
